more bugs fixed

+++ TransportGlobal/TransportGlobalAPI/src/TransportGlobal.Application/CQRSs/TransporterContextCQRSs/CommandDeleteEmployee/DeleteEmployeeCommandHandler.cs
@@ -29,6 +29,8 @@ namespace EventManagement.Application.CQRSs.LocationContextCQRSs.CommandDeleteLo
             UserEntity userEntity = _userRepository.GetByID(userID) ?? throw new ClientSideException(ExceptionConstants.NotFoundUser);
 
             EmployeeEntity employeeEntity = _employeeRepository.GetByID(request.ID) ?? throw new ClientSideException(ExceptionConstants.NotFoundEmployee);
+            if (employeeEntity.IsDeleted) throw new ClientSideException(ExceptionConstants.NotFoundEmployee);
+
             if (employeeEntity.CompanyID != userEntity.ActiveCompany?.ID) return Task.FromResult(new DeleteEmployeeCommandResponse(ResponseConstants.NotEmployeeOwner));
 
             employeeEntity.IsDeleted = true;
