+++ TransportGlobal/TransportGlobalAPI/src/TransportGlobal.Application/CQRSs/TransporterContextCQRSs/CommandDeleteVehicle/DeleteVehicleCommandHandler.cs
@@ -28,6 +28,8 @@ namespace TransportGlobal.Application.CQRSs.TransporterContextCQRSs.CommandDelet
             UserEntity userEntity = _userRepository.GetByID(userID) ?? throw new ClientSideException(ExceptionConstants.NotFoundUser);
 
             VehicleEntity vehicleEntity = _vehicleRepository.GetByID(request.ID) ?? throw new ClientSideException(ExceptionConstants.NotFoundVehicle);
+            if (vehicleEntity.IsDeleted) throw new ClientSideException(ExceptionConstants.NotFoundVehicle);
+
             if (vehicleEntity.CompanyID != userEntity.ActiveCompany?.ID) return Task.FromResult(new DeleteVehicleCommandResponse(ResponseConstants.NotVehicleOwner));
 
             if (vehicleEntity.Employees.Count > 0)
